+++ queries/bus.py
@@ -49,7 +49,13 @@ from islenska.basics import BinEntry
 import query
 from query import AnswerTuple, Query, QueryStateDict, ResponseType, Session
 from tree import Result, Node
-from queries import natlang_seq, cap_first, gen_answer, read_grammar_file, sing_or_plur
+from queries.util import (
+    natlang_seq,
+    cap_first,
+    gen_answer,
+    read_grammar_file,
+    sing_or_plur,
+)
 from speech.norm.num import floats_to_text, numbers_to_text
 from settings import Settings
 from reynir import correct_spaces
