+++ queries/counting.py
@@ -29,7 +29,7 @@ import logging
 import random
 from datetime import datetime, timedelta
 
-from queries import parse_num, gen_answer, read_grammar_file
+from queries.util import parse_num, gen_answer, read_grammar_file
 from query import Query, QueryStateDict
 from tree import Result, Node
