+++ queries/date.py
@@ -64,7 +64,7 @@ from datetime import datetime, timedelta
 from calendar import monthrange, isleap
 
 from query import Query, QueryStateDict
-from queries import (
+from queries.util import (
     gen_answer,
     is_plural,
     sing_or_plur,
