+++ queries/dictionary.py
@@ -30,7 +30,13 @@ import logging
 from query import Query, QueryStateDict
 from tree import Result, Node
 
-from queries import query_json_api, gen_answer, cap_first, icequote, read_grammar_file
+from queries.util import (
+    query_json_api,
+    gen_answer,
+    cap_first,
+    icequote,
+    read_grammar_file,
+)
 
 
 # This module wants to handle parse trees for queries
