+++ queries/geography.py
@@ -35,7 +35,7 @@ from datetime import datetime, timedelta
 from cityloc import capital_for_cc  # type: ignore
 
 from query import Query, QueryStateDict
-from queries import country_desc, nom2dat, cap_first, read_grammar_file
+from queries.util import country_desc, nom2dat, cap_first, read_grammar_file
 from reynir import NounPhrase
 from geo import (
     icelandic_city_name,
