+++ queries/pic.py
@@ -27,7 +27,7 @@ import logging
 from urllib.parse import urlparse
 
 from query import Query, QueryStateDict
-from queries import gen_answer, icequote, read_grammar_file
+from queries.util import gen_answer, icequote, read_grammar_file
 from reynir import NounPhrase
 from tree import Result, Node
 from images import get_image_url, Img
