+++ queries/places.py
@@ -36,11 +36,12 @@ import random
 from datetime import datetime
 
 from reynir import NounPhrase
-from iceaddr import nearest_addr, nearest_placenames
+
+# from iceaddr import nearest_addr, nearest_placenames
 
 from geo import in_iceland, iceprep_for_street, LatLonTuple
 from query import Query, QueryStateDict
-from queries import (
+from queries.util import (
     gen_answer,
     query_places_api,
     query_place_details,
