+++ queries/play.py
@@ -32,7 +32,7 @@ from random import choice
 from pyyoutube import Api, SearchListResponse
 
 from query import Query
-from queries import gen_answer
+from queries.util import gen_answer
 from utility import read_api_key
