+++ queries/repeat.py
@@ -27,7 +27,7 @@
 # TODO: Hvað spurði ég þig um síðast? Hvert var síðasta svar þitt? Endurtaktu þetta, o.s.frv.
 
 from query import Query
-from queries import gen_answer, icequote
+from queries.util import gen_answer, icequote
 from datetime import datetime, timedelta
