+++ queries/schedules.py
@@ -38,7 +38,7 @@ import datetime
 import cachetools
 
 from settings import changedlocale
-from queries import query_json_api, read_grammar_file
+from queries.util import query_json_api, read_grammar_file
 from tokenizer import split_into_sentences
