+++ queries/test.py
@@ -26,7 +26,7 @@
 """
 
 from query import Query
-from queries import gen_answer
+from queries.util import gen_answer
 
 
 _TEST_QTYPE = "Test"
