+++ queries/unit.py
@@ -35,7 +35,7 @@ import random
 from math import floor, log10
 
 from query import Query, QueryStateDict, to_dative, to_accusative
-from queries import iceformat_float, parse_num, read_grammar_file
+from queries.util import iceformat_float, parse_num, read_grammar_file
 from tree import Result, Node
