+++ queries/userinfo.py
@@ -34,7 +34,7 @@ from reynir.bindb import GreynirBin
 
 from geo import icelandic_addr_info, iceprep_for_placename, iceprep_for_street
 from query import ClientDataDict, Query
-from . import gen_answer
+from queries.util import gen_answer
 from speech.norm.num import numbers_to_text
