+++ queries/yulelads.py
@@ -30,7 +30,7 @@ from datetime import datetime
 
 from query import Query, QueryStateDict
 from tree import Result, Node, TerminalNode
-from queries import read_grammar_file
+from queries.util import read_grammar_file
 from speech.norm.num import numbers_to_ordinal
